refactor(#20): TokenManager 위치 변경 및 ReissueAuthenticator synchronized 추가
