+++ app/src/main/java/com/ssafy/neegongnaegong/data/remote/utils/TokenManager.kt
@@ -1,4 +1,4 @@
-package com.ssafy.neegongnaegong.data.remote
+package com.ssafy.neegongnaegong.data.remote.utils
 
 interface TokenManager {
     fun saveToken(tokenType: String, token: String)
